fix: replace expence-recurring[ref:#1862]
