+++ apps/gauzy/src/app/pages/expense-recurring/expense-recurring-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+import { ExpenseRecurringComponent } from './expense-recurring.component';
+
+const routes: Routes = [
+	{
+		path: '',
+		component: ExpenseRecurringComponent
+	}
+];
+
+@NgModule({
+	imports: [RouterModule.forChild(routes)],
+	exports: [RouterModule]
+})
+export class ExpenseRecurringRoutingModule {}
