+++ apps/gauzy/src/app/pages/expense-recurring/expense-recurring.module.ts
@@ -1,0 +1,44 @@
+import { NgModule } from '@angular/core';
+import { ThemeModule } from '../../@theme/theme.module';
+import {
+	NbCardModule,
+	NbButtonModule,
+	NbInputModule,
+	NbIconModule,
+	NbDialogModule
+} from '@nebular/theme';
+import { FormsModule } from '@angular/forms';
+import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
+import { HttpClient } from '@angular/common/http';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { ExpenseRecurringRoutingModule } from './expense-recurring-routing.module';
+import { ExpenseRecurringComponent } from './expense-recurring.component';
+import { RecurringExpenseBlockModule } from '../../@shared/expenses/recurring-expense-block/recurring-expense-block.module';
+
+export function HttpLoaderFactory(http: HttpClient) {
+	return new TranslateHttpLoader(http, './assets/i18n/', '.json');
+}
+
+@NgModule({
+	imports: [
+		ExpenseRecurringRoutingModule,
+		ThemeModule,
+		NbCardModule,
+		FormsModule,
+		NbButtonModule,
+		NbInputModule,
+		NbIconModule,
+		NbDialogModule.forChild(),
+		TranslateModule.forChild({
+			loader: {
+				provide: TranslateLoader,
+				useFactory: HttpLoaderFactory,
+				deps: [HttpClient]
+			}
+		}),
+		RecurringExpenseBlockModule
+	],
+	declarations: [ExpenseRecurringComponent],
+	entryComponents: []
+})
+export class ExpenseRecurringModule {}
