+++ apps/gauzy/src/app/pages/organizations/edit-organization/edit-organization.component.scss
@@ -1,61 +1,21 @@
 .edit-icon {
-	margin-left: 30px;
-	position: relative;
-	width: 36px;
+  margin-left: 30px;
+  position: relative;
+  width: 36px;
 
-	svg {
-		position: absolute;
-	}
+  svg {
+    position: absolute;
+  }
 
-	nb-icon {
-		position: absolute;
-	}
+  nb-icon {
+    position: absolute;
+  }
 }
 
 .org-details {
-	.edit-public-page {
-		cursor: pointer;
-		color: #027ad6;
-		padding-top: 3px;
-	}
-}
-
-.setting-name {
-	font-size: 24px;
-	font-weight: bold;
-}
-
-.body-header {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	margin-bottom: 35px;
-}
-
-.mutation-card.setting-block {
-	background: #eaf3fc;
-}
-
-.transparent {
-	opacity: 0.7;
-}
-
-.settings-body {
-	padding: 35px;
-}
-
-.sub-header {
-	margin-bottom: 20px;
-	font-weight: bold;
-}
-
-.header-content {
-	display: flex;
-	.header-info {
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-		width: 560px;
-		padding-left: 30px;
-	}
+  .edit-public-page {
+    cursor: pointer;
+    color: #027ad6;
+    padding-top: 3px;
+  }
 }
